Review types

+++ src/channel.rs
@@ -8,9 +8,9 @@ use std::io::Cursor;
 pub struct ChannelConfig {
     name: String,
     typ: String,
-    shape: Option<Vec<i32>>,
+    shape: Option<Vec<u32>>,
     elements: usize,
-    element_size: usize,
+    element_size: u32,
     le: bool,
     compression: String,
 }
@@ -22,7 +22,7 @@ impl ChannelConfig {
     pub fn get_type(&self) -> String {
         self.typ.clone()
     }
-    pub fn get_shape(&self) -> Option<Vec<i32>> {
+    pub fn get_shape(&self) -> Option<Vec<u32>> {
         self.shape.clone()
     }
     pub fn get_elements(&self) -> usize {
@@ -31,7 +31,7 @@ impl ChannelConfig {
     pub fn get_compression(&self) -> String {
         self.compression.clone()
     }
-    pub fn get_element_size(&self) -> usize {
+    pub fn get_element_size(&self) -> u32 {
         self.element_size.clone()
     }
 }
@@ -47,7 +47,7 @@ pub struct ChannelArray<T> {
     buffer: Vec<T>,
 }
 
-pub fn get_elements(shape: &Option<Vec<i32>>) -> usize {
+pub fn get_elements(shape: &Option<Vec<u32>>) -> usize {
     let nelm = shape.clone()
         .filter(|v| !v.is_empty()) // Ensure it's not empty
         .map(|v| v.into_iter().product()) // Compute product of elements
@@ -56,7 +56,7 @@ pub fn get_elements(shape: &Option<Vec<i32>>) -> usize {
     elements
 }
 
-fn get_element_size(typ: &str) -> usize {
+fn get_element_size(typ: &str) -> u32 {
     match typ {
         "bool" => 4,
         "string" => 1,
@@ -74,7 +74,7 @@ fn get_element_size(typ: &str) -> usize {
     }
 }
 impl<T: Default + Clone> ChannelScalar<T> {
-    pub fn new(name: String, typ: String, shape: Option<Vec<i32>>, le: bool, compression: String, reader: fn(&mut Cursor<&Vec<u8>>) -> IOResult<T>) -> Self {
+    pub fn new(name: String, typ: String, shape: Option<Vec<u32>>, le: bool, compression: String, reader: fn(&mut Cursor<&Vec<u8>>) -> IOResult<T>) -> Self {
         let elements = get_elements(&shape);
         let element_size = get_element_size(&typ);
         let config = ChannelConfig { name, typ, shape, elements, element_size, le, compression };
@@ -84,7 +84,7 @@ impl<T: Default + Clone> ChannelScalar<T> {
 
 
 impl<T: Default + Clone> ChannelArray<T> {
-    pub fn new(name: String, typ: String, shape: Option<Vec<i32>>, le: bool, compression: String, reader: fn(&mut Cursor<&Vec<u8>>, &mut [T]) -> IOResult<()>) -> Self {
+    pub fn new(name: String, typ: String, shape: Option<Vec<u32>>, le: bool, compression: String, reader: fn(&mut Cursor<&Vec<u8>>, &mut [T]) -> IOResult<()>) -> Self {
         let elements = get_elements(&shape);
         let element_size = get_element_size(&typ);
         let config = ChannelConfig { name, typ, shape, elements, element_size, le, compression };
